backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.0.1
fastapi==0.103.1
pydantic==2.3.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pymongo
uvicorn==0.23.2